Refactor release notes to be by API and to have tighter structure (#357)
* Refactor release notes to be by API and to have tighter structure

* Remove src/content/releaseNotes directory and update release notes doc

* Extract some of the apiDocs content into category index files

* Index all content at category level and omit removed APIs from release notes
