+++ src/apiDefs/data/appeals.ts
@@ -1,3 +1,4 @@
+import { AppealsReleaseNotes } from '../../content/apiDocs/appeals';
 import { IApiDescription } from '../schema';
 
 const swaggerHost : string = process.env.REACT_APP_VETSGOV_SWAGGER_API!;
@@ -12,6 +13,7 @@ const appealsApis : IApiDescription[] = [
     ],
     enabledByDefault: false,
     name: 'Decision Reviews API',
+    releaseNotes: AppealsReleaseNotes,
     trustedPartnerOnly: false,
     urlFragment: 'decision_reviews',
     vaInternalOnly: true,
