+++ src/apiDefs/deprecated.test.ts
@@ -1,5 +1,8 @@
 import 'jest';
 import * as moment from 'moment';
+
+// we just need a Markdown component for  our test IApiDescription
+import { FhirApiReleaseNotes } from '../content/apiDocs/health';
 import { IApiDescription } from './schema';
 
 jest.mock('./query');
@@ -16,6 +19,7 @@ describe('deprecated API module', () => {
       enabledByDefault: true,
       name: 'My API',
       oAuth: false,
+      releaseNotes: FhirApiReleaseNotes,
       trustedPartnerOnly: false,
       urlFragment: 'my_api',
       vaInternalOnly: false,
