+++ src/containers/releaseNotes/CategoryReleaseNotesPage.tsx
@@ -12,14 +12,26 @@ import PageHeader from '../../components/PageHeader';
 import { defaultFlexContainer } from '../../styles/vadsUtils';
 import { IApiNameParam } from '../../types';
 
+const ApiReleaseNote = ({ api }: { api: IApiDescription }) => {
+  const dashUrlFragment = api.urlFragment.replace('_', '-');
+  return (
+    <Flag name={`hosted_apis.${api.urlFragment}`}>
+      <div id={dashUrlFragment}>
+        <h2>{api.name}</h2>
+        {api.releaseNotes({})}
+        <hr/>
+      </div>
+    </Flag>
+  );
+};
+
 export default class CategoryReleaseNotesPage extends React.Component<
-  RouteComponentProps<IApiNameParam>,
-  {}
+  RouteComponentProps<IApiNameParam>
 > {
   public render() {
     const apiDefs = getApiDefinitions();
     const { apiCategoryKey } = this.props.match.params;
-    const { apis, content } = apiDefs[apiCategoryKey];
+    const { apis } = apiDefs[apiCategoryKey];
 
     const headerProps = {
       halo: 'Release Notes',
@@ -64,9 +76,8 @@ export default class CategoryReleaseNotesPage extends React.Component<
         <PageHeader halo={headerProps.halo} header={headerProps.header} />
         {cardSection}
         <div className={classNames('vads-u-width--full', 'vads-u-margin-top--4')}>
-          {content.releaseNotes ? content.releaseNotes({}) : null}
+          {apis.map(api => <ApiReleaseNote key={api.urlFragment} api={api} />)}
         </div>
-        <hr />
       </section>
     );
   }
