+++ src/content/apiDocs/appeals/appealsReleaseNotes.mdx
@@ -1,7 +1,2 @@
-<div id="appeals" />
-
-# Decision Reviews
-## Month 99, 2099
+### Month 99, 2099
 (v1) These are dummy release notes.  Link to PRs here!
-
----
