+++ src/content/apiDocs/benefits/appealsReleaseNotes.mdx
@@ -1,0 +1,17 @@
+### January 31, 2019
+Add Power of Attorney verification for future Appeals Version 1 [#2717](https://github.com/department-of-veterans-affairs/vets-api/pull/2717)
+
+---
+
+### January 14, 2019
+Bug discovered and fixed for safe null requests [#2727](https://github.com/department-of-veterans-affairs/vets-api/pull/2727)
+
+---
+
+### June 11, 2018
+Implement VA-required headers to retrieve status [#2024](https://github.com/department-of-veterans-affairs/vets-api/pull/2024)
+
+---
+
+### May 23, 2018
+Add appeals service API [#1961](https://github.com/department-of-veterans-affairs/vets-api/pull/1961)
