+++ src/content/apiDocs/benefits/benefitsIntakeReleaseNotes.mdx
@@ -1,0 +1,131 @@
+### October 7th, 2019
+(v0 & v1) Add support for base64 encoding [#3995](https://github.com/department-of-veterans-affairs/vets-api/pull/3395)
+
+---
+
+### July 29, 2019
+(v0 & v1) New `expired` status added to individual status endpoint [#3353](https://github.com/department-of-veterans-affairs/vets-api/pull/3353)
+
+---
+
+### August 19, 2019
+(v0 & v1) Validation on document page size [#3213](https://github.com/department-of-veterans-affairs/vets-api/pull/3213)
+
+---
+
+### July 24, 2019
+(v0 & v1) Cache statuses for /uploads/report endpoint [#3158](https://github.com/department-of-veterans-affairs/vets-api/pull/3158)
+
+---
+
+### July 11th, 2019
+(v1) Revert changes to error handling [#3150](https://github.com/department-of-veterans-affairs/vets-api/pull/3150)
+
+---
+
+### July 9th, 2019
+(v1) Announced support for v1, which includes 2 breaking changes:
+* "vbms" status [#3027](https://github.com/department-of-veterans-affairs/vets-api/pull/3027)
+* Updates to error handling [#3053](https://github.com/department-of-veterans-affairs/vets-api/pull/3053)
+
+---
+
+### June 11, 2019
+(v0 & v1) Adding Metadata Endpoints for V0 and V1 end points [#3093](https://github.com/department-of-veterans-affairs/vets-api/pull/3093)
+
+---
+
+### June 3, 2019
+(v0) Upgrade Ruby to 2.4.5 [#2995](https://github.com/department-of-veterans-affairs/vets-api/pull/2995)
+
+---
+
+### May 28, 2019
+(v0) Rails 5.2.3 [#2933](https://github.com/department-of-veterans-affairs/vets-api/pull/2933)
+
+---
+
+### May 23, 2019
+(v0) Create self-service testing tool [#3059](https://github.com/department-of-veterans-affairs/vets-api/pull/3059)
+
+---
+
+### May 3, 2019
+(v0) Validate Against Empty Files [#3025](https://github.com/department-of-veterans-affairs/vets-api/pull/3025)
+
+---
+
+### May 1, 2019
+(v0) Rails 5.1 [#2897](https://github.com/department-of-veterans-affairs/vets-api/pull/2897)
+
+---
+
+### April 9, 2019
+(v0) Upgrade API to Rails 5.0.7.2 [#2833](Upgrade API to Rails 5.0.7.2 (#2833)
+
+---
+
+### March 18, 2019
+(v0) Handle new vbms success status without passing it along to consumers (would be a breaking change) [#2877](https://github.com/department-of-veterans-affairs/vets-api/pull/2877)
+
+---
+
+### January 28, 2019
+(v0) Status simulation in Developer environment [#2751](https://github.com/department-of-veterans-affairs/vets-api/pull/2751)
+- Can now specify status in developer environment, and receive mocked data for that status.
+
+---
+
+### October 8, 2018
+(v0) Update API URLs to va.gov domains in documentation [#2332](https://github.com/department-of-veterans-affairs/vets-api/pull/2332)
+
+---
+
+### September 18, 2018
+(v0) Implement VA-required logging for uploads [#2072](https://github.com/department-of-veterans-affairs/vets-api/pull/2072)
+
+---
+
+### August 27, 2018
+(v0) Validating PDF format on upload [#2213](https://github.com/department-of-veterans-affairs/vets-api/pull/2213)
+
+---
+
+### July 23, 2018
+(v0) Limit the size of individual attachments in uploads [#2094](https://github.com/department-of-veterans-affairs/vets-api/pull/2094)
+
+---
+
+### July 19, 2018
+* (v0) Improve pension and burial claim uploads to Central Mail [#2143](https://github.com/department-of-veterans-affairs/vets-api/pull/2143)
+* (v0) Add Consumer ID to Central Mail upload submission [#2071](https://github.com/department-of-veterans-affairs/vets-api/pull/2071)
+
+---
+
+### July 2, 2018
+(v0) Add Amazon SNS support to endpoint [#2058](https://github.com/department-of-veterans-affairs/vets-api/pull/2058)
+
+---
+
+### June 8, 2018
+(v0) Add numeric-characters-only validation to fileNumber metadata element [#2020](https://github.com/department-of-veterans-affairs/vets-api/pull/2020)
+
+---
+
+### May 17, 2018
+(v0) Add bulk status report [#1947](https://github.com/department-of-veterans-affairs/vets-api/pull/1947)
+
+---
+
+### May 8, 2018
+(v0) Start uploading documents to Central Mail [#1896](https://github.com/department-of-veterans-affairs/vets-api/pull/1896)
+
+---
+
+### April 26, 2018
+(v0) Start retrieving statuses from Central Mail [#1887](https://github.com/department-of-veterans-affairs/vets-api/pull/1887)
+
+---
+
+### April 20, 2018
+(v0) Create initial service [#1887](https://github.com/department-of-veterans-affairs/vets-api/pull/1887)
