+++ src/content/apiDocs/benefits/claimsReleaseNotes.mdx
@@ -1,0 +1,148 @@
+### November 27, 2019
+(v0) Add "claimDate" field to set receipt dates for internal consumers [#3596](https://github.com/department-of-veterans-affairs/vets-api/pull/3596)
+
+---
+
+### November 20, 2019
+(v1) Require attachment for original claims submitted by representatives [#3554](https://github.com/department-of-veterans-affairs/vets-api/pull/3554)
+
+---
+
+### November 12, 2019
+(v1) Validate existence of previous claim to determine "original claim" status and require attachment in those cases [#3471](https://github.com/department-of-veterans-affairs/vets-api/pull/3471)
+
+---
+
+### September 24, 2019
+(v0 & v1) Add "autoCestPDFGenerationDisabled" boolean field [#3328](https://github.com/department-of-veterans-affairs/vets-api/pull/3328)
+
+---
+
+### July 23, 2019
+(v1) Announced production support for external use of v1, allowing:
+* Users with existing Power of Attorney to establish 526 claims on a Veteran's behalf
+* Authenticated Veterans to establish 526 claims on their own behalf
+
+---
+
+### June 25, 2019
+Creating ITF expiration validation for 526 [#3119](https://github.com/department-of-veterans-affairs/vets-api/pull/3119)
+
+---
+
+### June 17, 2019
+Add supporting documents to claims [#3111](https://github.com/department-of-veterans-affairs/vets-api/pull/3111)
+
+---
+
+### June 10, 2019
+Add machine readable source to errors [#3097](https://github.com/department-of-veterans-affairs/vets-api/pull/3097)
+
+---
+
+### June 3, 2019
+* Add a couple specs for empty ITF requests [#3083](https://github.com/department-of-veterans-affairs/vets-api/pull/3083)
+* Upgrade Ruby to 2.4.5 [#2995](https://github.com/department-of-veterans-affairs/vets-api/pull/2995)
+
+---
+
+### May 30, 2019
+* Adding logic to match claims detail and auto claim serializers [#3061](https://github.com/department-of-veterans-affairs/vets-api/pull/3061)
+* Adding Form Schema documentation and json api compliance [#3075](https://github.com/department-of-veterans-affairs/vets-api/pull/3075)
+
+---
+
+### May 28, 2019
+* Rails 5.2.3 [#2933](https://github.com/department-of-veterans-affairs/vets-api/pull/2933)
+* Get request for displaying JSON Schema for each form [#3072](https://github.com/department-of-veterans-affairs/vets-api/pull/3072)
+
+---
+
+### May 20, 2019
+* (v1) Claims docs fix [#3055](https://github.com/department-of-veterans-affairs/vets-api/pull/3055)
+* (v1) add edipi to from_identity call [#3047](https://github.com/department-of-veterans-affairs/vets-api/pull/3047)
+
+---
+
+### May 1, 2019
+* (v1) Art.claims 404 [#3015](https://github.com/department-of-veterans-affairs/vets-api/pull/3015)
+* (v1) Rails 5.1 [#2897](https://github.com/department-of-veterans-affairs/vets-api/pull/2897)
+
+---
+
+### April 30, 2019
+(v1) Creating Swagger Blocks Documentation for Claims V1 [#2952](https://github.com/department-of-veterans-affairs/vets-api/pull/2952)
+
+---
+
+### April 23, 2019
+* (v1) Check PoA with Okta [#2990](https://github.com/department-of-veterans-affairs/vets-api/pull/2990)
+* (v1) Intent To File Claims Api [#2990](https://github.com/department-of-veterans-affairs/vets-api/pull/2990)
+
+---
+
+### April 17, 2019
+(v1) Art.claims api schema [#2965](https://github.com/department-of-veterans-affairs/vets-api/pull/2965)
+
+---
+
+### April 12, 2019
+(v1) Verify PoA from OGC data [#2951](https://github.com/department-of-veterans-affairs/vets-api/pull/2951)
+
+---
+
+### April 11, 2019
+(v1) Add ability to set users from oauth token or headers [#2949](https://github.com/department-of-veterans-affairs/vets-api/pull/2949)
+
+---
+
+### April 9, 2019
+* (v1) Upgrade API to Rails 5.0.7.2 [#2833](https://github.com/department-of-veterans-affairs/vets-api/pull/2833)
+* (v1 )Uploading Uploads to EVSS for Auto-CEST [#2915](https://github.com/department-of-veterans-affairs/vets-api/pull/2915)
+* (v1) 526 Additional Validations [#2936](https://github.com/department-of-veterans-affairs/vets-api/pull/2936)
+
+---
+
+### April 5, 2019
+(v1) Split off Claims v1 [#2927](https://github.com/department-of-veterans-affairs/vets-api/pull/2927)
+
+---
+
+### March 26, 2019
+(v0) Create mock override to test auto-establishment in staging [#2908](https://github.com/department-of-veterans-affairs/vets-api/pull/2908)
+- Our upstream partners limit the number of test submissions allowed in staging, so we now use mocked data by default.
+
+---
+
+### March 21, 2019
+(v0) Fix security vulnerability in disability claim auto-establishment [#2903](https://github.com/department-of-veterans-affairs/vets-api/pull/2903)
+
+---
+
+### March 20, 2019
+(v0) Fix duplicate submissions for disability claim auto-establishment [#2852](https://github.com/department-of-veterans-affairs/vets-api/pull/2852)
+
+---
+
+### March 18, 2019
+(v0) Add support for uploading documents to auto-established disability compensation claims [#2868](https://github.com/department-of-veterans-affairs/vets-api/pull/2868)
+
+---
+
+### March 15, 2019
+(v0) Improve Veteran lookup for disability claim auto-establishment and claim status [#2863](https://github.com/department-of-veterans-affairs/vets-api/pull/2863)
+
+---
+
+### January 8, 2019
+(v0) Implement initial VA internal-only disability claim auto-establishment [#2824](https://github.com/department-of-veterans-affairs/vets-api/pull/2824)
+
+---
+
+### January 8, 2019
+(v0) Implement Power of Attorney lookup for future v1 [#2709](https://github.com/department-of-veterans-affairs/vets-api/pull/2709)
+
+---
+
+### January 4, 2019
+(v0) Update claim statuses to be consistent with VA.gov [#2701](https://github.com/department-of-veterans-affairs/vets-api/pull/2701)
