+++ src/content/apiDocs/health/argonautReleaseNotes.mdx
@@ -1,0 +1,3 @@
+### October 1, 2019
+
+The Argonaut API was removed and replaced by the [FHIR Veterans Health API](/explore/health/docs/fhir)
