+++ src/content/apiDocs/health/fhirAPIReleaseNotes.mdx
@@ -1,0 +1,3 @@
+### December 4, 2018
+Launch of the Veterans Health API
+- Available for development environment access.
