--- src/content/apiDocs/health/healthReleaseNotes.mdx
@@ -1,6 +1,0 @@
-<div id="health" />
-
-## December 4, 2018
-Launch of the Veterans Health API
-- Available for development environment access.
-
