+++ src/content/apiDocs/vaForms/vaFormsReleaseNotes.mdx
@@ -1,6 +1,2 @@
-<div id="vaForms" />
-
-# VA Forms
-
-## November 20, 2019
+### November 20, 2019
 Initial release of the VA Forms API
