+++ src/content/apiDocs/verification/addressValidationReleaseNotes.mdx
@@ -1,0 +1,3 @@
+### September 26, 2018
+Create Address Validation API
+- The ‘validate’ API accepts a structured JSON address object (broken down by street/city/state/zip/etc), and returns a result indicating whether the address is valid, and if so includes a canonicalized address and geocoding information (lat/long) for the address.
