+++ src/content/apiDocs/verification/veteranConfirmationReleaseNotes.mdx
@@ -1,0 +1,5 @@
+### January 13, 2020
+Launched Veteran Confirmation API with a status endpoint
+- `/status`: Allows third-parties to request confirmation from the VA of an individual's Veteran status after providing a valid API key and identifying attributes for the individual.
+
+[View code change(s)](https://github.com/department-of-veterans-affairs/vets-api/pull/3676)
