--- src/content/releaseNotes/release-notes.mdx
@@ -1,82 +1,0 @@
-# Release Notes
-
-Important information about API updates.
-
-<br />
-<div id="benefits" />
-
-## Benefits
-
-### Benefits Intake API
-
-**November 2018 | v0.0.0**
-Launched v0.0.0 of the Benefits Intake API.
-
-<br />
-
-### Appeals Status API
-
-**November 2018 | v0.0.1**
-Launched v0.0.1 of the Appeals API.
-
-<br />
-
-### Benefits Claims API
-
-**November 2018 | v0.0.1**
-Launched v0.0.1 of the Benefits Claims API.
-
-<br />
-<div id="facilities" />
-
-## Facilities
-
-### VA Facilities API
-
-**January 2019 | v0.0.1**
-Both a list of IDs and a lat/long can now be passed in the same request.
-
-**August 2018 | v0.0.1**
-Launched v0.0.1 of the VA Facilities API.
-
-<br />
-<div id="health" />
-
-## Health
-
-### Veterans Health API
-
-**November 2018 | v1**
-Launched v1 of the Argonaut API.
-
-<br />
-<div id="verification" />
-
-## Veteran Verification
-
-### Disability Rating
-
-**November 2018 | v0.0.1**
-Launched v0.0.1 of the Disability Rating API.
-
-<br />
-
-### Service History
-
-**November 2018 | v0.0.1**
-Launched v0.0.1 of the Service History API.
-
-<br />
-
-### Veteran Confirmation
-
-**November 2018 | v0.0.1**
-Launched v0.0.1 of the Veteran Confirmation API.
-
-<br />
-
-### Address Validation
-
-**November 2018 | v1**
-Launched v1 of the Address Validation API.
-
